should fix unit tests for gh actions
